Django_api/requirements.txt
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
django-filter
mysqlclient